Separate the aggregation code into a module of its own.
