+++ examples/satellites/check_stability.py
@@ -18,123 +18,18 @@ import matplotlib
 matplotlib.use("Agg")
 import matplotlib.pylab as plt
 
-import contextlib
 import math
 import os
 import re
 import string
-import time
 import warnings
 
 import pandas as pd
 import seaborn as sns
 
-import bayeslite
-
-######################################################################
-## Query running and result accumulation                            ##
-######################################################################
-
-model_schedule = [1,3,6]
-model_skip = model_schedule[-1]
-n_replications = 10
-# TODO Expect the number of models in the file to be at least
-# model_skip * n_replications; excess models are wasted.
-
-then = time.time()
-def log(msg):
-    print "At %3.2fs" % (time.time() - then), msg
-
-def analyze_fileset(files):
-    """Aggregate all the queries over all the given bdb files.
-
-    Do seed and model count variation within each file by varying
-    model specs; assume the files represent analysis iteration
-    variation and report accordingly.
-
-    Alternative: could just assume the file name describes the file
-    opaquely.
-
-    """
-    # Keys are (file_name, model_ct, name); values are aggregated results
-    results = {}
-    for fname in files:
-        log("processing file %s" % fname)
-        with bayeslite.bayesdb_open(fname) as bdb:
-            res = [((fname, model_ct, name), ress)
-                   for ((model_ct, name), ress)
-                   in analyze_queries(bdb).iteritems()]
-            incorporate(results, res)
-    return results
-
-def model_specs():
-    def spec_at(location, size):
-        """Return a 0-indexed model range
-        inclusive of lower bound and exclusive of upper bound."""
-        return (location * model_skip, location * model_skip + size)
-    return [spec_at(location, size)
-            for location in range(n_replications)
-            for size in model_schedule]
-
-@contextlib.contextmanager
-def model_restriction(bdb, gen_name, spec):
-    (low, high) = spec
-    model_count = 60 # TODO !
-    with bdb.savepoint_rollback():
-        if low > 0:
-            bdb.execute('''DROP MODELS 0-%d FROM %s''' % (low-1, gen_name))
-        if model_count > high:
-            bdb.execute('''DROP MODELS %d-%d FROM %s'''
-                        % (high, model_count-1, gen_name))
-        yield
-
-def incorporate(running, new):
-    for (key, more) in new:
-        if key in running:
-            running[key] = merge_one(running[key], more)
-        else:
-            running[key] = more
-
-# Type-dependent aggregation of intermediate values (e.g., counting
-# yes and no for booleans vs accumulating lists of reals).
-def inc_singleton(qtype, val):
-    if qtype == 'num':
-        return ('num', [val])
-    elif qtype == 'bool':
-        if val:
-            return ('bool', (1, 0))
-        else:
-            return ('bool', (0, 1))
-    else:
-        raise Exception("Unknown singleton type %s for %s" % (qtype, val))
-
-def merge_one(so_far, val):
-    (tp1, item1) = so_far
-    (tp2, item2) = val
-    if tp1 == 'num' and tp2 == 'num':
-        item1.extend(item2)
-        return ('num', item1)
-    elif tp1 == 'bool' and tp2 == 'bool':
-        (t1, f1) = item1
-        (t2, f2) = item2
-        return ('bool', (t1+t2, f1+f2))
-    else:
-        raise Exception("Type mismatch in merge into %s of %s" % (so_far, val))
-
-def analyze_queries(bdb):
-    results = {} # Keys are (model_ct, name); values are aggregated results
-    for spec in model_specs():
-        (low, high) = spec
-        model_ct = high - low
-        with model_restriction(bdb, "satellites_cc", spec):
-            log("querying models %d-%d" % (low, high-1))
-            for queryset in [country_purpose_queries,
-                             unlikely_periods_queries,
-                             orbit_type_imputation_queries]:
-                qres = [((model_ct, name), inc_singleton(qtype, res))
-                        for (name, qtype, res) in queryset(bdb)]
-                incorporate(results, qres)
-    return results
+import sys
+sys.path.append("..")
+from aggregation import log, analyze_fileset
 
 ######################################################################
 ## Visualization                                                    ##
@@ -344,10 +239,12 @@ import cPickle as pickle # json doesn't like tuple dict keys
 import glob
 
 def save_query_results(filename):
-    files = glob.glob("output/*i.bdb")
-    # files = ["output/satellites-2015-09-30-axch-60m-4i.bdb",
-    #          "output/satellites-2015-09-30-axch-60m-8i.bdb"]
-    results = analyze_fileset(files)
+    # files = glob.glob("output/*i.bdb")
+    files = ["output/satellites-2015-09-30-axch-60m-4i.bdb",
+             "output/satellites-2015-09-30-axch-60m-8i.bdb"]
+    results = analyze_fileset(files, [country_purpose_queries,
+                                      unlikely_periods_queries,
+                                      orbit_type_imputation_queries])
 
     with open(filename, "w") as f:
         pickle.dump(results, f)
